add configuration to set custom field extractor
