+++ src/main/scala/com/unstablebuild/slime/Encoder.scala
@@ -1,13 +1,9 @@
 package com.unstablebuild.slime
 
-import java.time.{Instant, ZoneOffset}
-import java.time.format.DateTimeFormatter
-
 import ch.qos.logback.classic.spi.LoggingEvent
 import ch.qos.logback.core.encoder.EncoderBase
+import com.unstablebuild.slime.field.{DefaultFieldExtractor, FieldExtractor}
 import com.unstablebuild.slime.format.Text
-
-import scala.collection.JavaConverters._
 
 class Encoder extends EncoderBase[LoggingEvent] {
 
@@ -17,7 +13,7 @@ class Encoder extends EncoderBase[LoggingEvent] {
 
   var fields: Seq[String] = Seq("level", "message")
 
-  var fieldExtractors: Map[String, LoggingEvent => Value] = Encoder.fieldExtractors
+  var fieldExtractor: FieldExtractor = new DefaultFieldExtractor
 
   override def encode(event: LoggingEvent): Array[Byte] = {
     doDebug(s"encode $event")
@@ -39,7 +35,7 @@ class Encoder extends EncoderBase[LoggingEvent] {
   }
 
   def extractFields(event: LoggingEvent): Seq[(String, Value)] = {
-    fields.flatMap(f => Encoder.fieldExtractors.get(f).map(extract => f -> extract(event)))
+    fields.flatMap(f => fieldExtractor.extract(f, event).map(f -> _))
   }
 
   override def headerBytes(): Array[Byte] = {
@@ -60,26 +56,13 @@ class Encoder extends EncoderBase[LoggingEvent] {
     this.fields = allFields.split(",").map(_.toLowerCase.trim)
   }
 
+  def setFieldExtractor(fieldExtractor: FieldExtractor): Unit = {
+    this.fieldExtractor = fieldExtractor
+  }
+
   @inline
   private def doDebug(msg: => String): Unit = {
     if (debug) addInfo(msg)
   }
-
-}
-
-object Encoder {
-
-  private val dateFormatter = DateTimeFormatter.ISO_ZONED_DATE_TIME.withZone(ZoneOffset.UTC)
-
-  val fieldExtractors: Map[String, LoggingEvent => Value] =
-    Map(
-      "level" -> (e => StringValue(e.getLevel.toString)),
-      "message" -> (e => StringValue(e.getFormattedMessage)),
-      "thread" -> (e => StringValue(e.getThreadName)),
-      "logger" -> (e => StringValue(e.getLoggerName)),
-      "mdc" -> (e => NestedValue(e.getMDCPropertyMap.asScala.mapValues(StringValue).toSeq)),
-      "timestamp" -> (e => NumberValue(e.getTimeStamp)),
-      "ts" -> (e => StringValue(dateFormatter.format(Instant.ofEpochSecond(e.getTimeStamp))))
-    )
 
 }
