+++ src/test/resources/logback-test.xml
@@ -12,7 +12,7 @@
   <appender name="console" class="ch.qos.logback.core.ConsoleAppender">
     <encoder class="com.unstablebuild.slime.Encoder">
       <format class="com.unstablebuild.slime.format.ColoredText" />
-      <fields>level,logger,thread,message</fields>
+      <fields>level,ts,caller,logger,thread,message</fields>
     </encoder>
   </appender>
   <root level="all">
